requirements.txt
asgiref==3.4.1
autopep8==1.6.0
Django==4.0
django-axes==5.24.0
django-cors-headers==3.10.1
django-ipware==4.0.0
djangorestframework==3.12.4
gunicorn==20.1.0
Pillow==8.4.0
psycopg2-binary==2.9.1
pycodestyle==2.8.0
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
whitenoise==5.3.0
